requirements.txt
streamlit==1.31.0
cryptography==41.0.0
Pillow==10.2.0
python-dateutil==2.8.2
requests==2.31.0
caldav==1.3.6
icalendar==5.0.10